+++ CHANGELOG.md
@@ -1,6 +1,66 @@
-### Unreleased
+## Unreleased
 
-### Release
+### 4.1.0
+#### Minor Change
+##### Bookmark List and Bookmark Widget refactoring
+- Server side rendering and code refactoring of BookmarkList and BookmarkWidget
+
+##### Article Full Screen Layout Adjustment
+- Update src/managers/layout-manager.js. Render different layout for `article:fullscreen:[normal|dark]` theme
+- Update src/containers/Article.js
+  - Stop using `post.theme` to  render article layout.
+  - Render different article layout according to `post.style`
+  - Add `isLeadingAssetFullScreen` and `styles` React Props
+- Update src/components/article/layout/title-row-[above|upon].js
+- Change prop `theme` to `styles` and do some design tuning
+- Update src/components/article/layout/layout-maker.js.
+  - Replace `position.title` and `position.header` by `isLeadingAssetFullScreen`
+  - Render full screen image with a gray mask
+  - Replace `theme` by `styles`
+
+##### Dependencies Upgrade
+- redux-thunk to^2.3.0
+- @twreporter/universal-header to ^2.0.3
+
+##### Miscellaneous
+- Turn off webpack progress display
+- Check if service worker registration exists before using it
+
+## Release
+### 4.0.1
+- Remove props `history` from `BrowserRouter`.
+
+### 4.0.0
+#### Major change
+##### Dependencies Upgrade
+- react, react-dom to ^16.3.0
+- styled-components to ^4.0.0
+- react-router-dom to ^4.0.0
+- react-redux to ^6.0.0
+- @twreporter/react-components to ^6.1.0
+- @twreporter/universal-header to ^2.0.0
+- drop @twreporter/registration
+
+##### Introduce new Authentication/Authorization API
+- check/grant auth on server side through express auth middleware
+- store auth info into redux store
+- enable server side rendering for personal requests
+
+##### Render Universal Header with @twreporter/universal-header
+- universal header for personal requests with server side rendering
+
+##### Render react-router-dom v4 routes on server side with react-loadable
+- add `src/data-loaders` folder, which lists data loading functions for different pages.
+- implement react-loadabel server side rendering since react-router-dom v4 doesn't support
+`getComponent` callback function anymore.
+
+##### Introduce layout and theme manager to render page layout and theme
+- deal with page layout and theme in `src/containers/app-shell.js`
+- drop HOC layout helper(`src/helpers/with-layout.js`)
+
+##### Miscellaneous
+- Move category related constants into src/constants/category.js
+- Update src/containers/app-shell.js. Refactor web push registration promise chain
 
 ### 3.2.13
 - [Bug] Add a workaround to author data problem